arrangement api structure
